+++ js/data.js
@@ -1,0 +1,42 @@
+const DESCRIPTIONS = [
+  'Пляж в бухте',
+  'Заморский салат',
+  'Бикини фотограф',
+  'Тапочки с подсветкой',
+  'Ламба-бомба',
+  'Обувница',
+  'Портрет краба',
+  'Хор мальчиков',
+  'Самолет в небе',
+  'Салат с орехами',
+  'Суши-Кот',
+  'Тачка на прокачку',
+];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+const NAMES = [
+  'Вася Пупкин',
+  'Денис Шведов',
+  'Джиа Мари Каранджи',
+  'Бульсинея Табольская',
+  'Клеопатра',
+  'Каринка Мандаринка',
+  'Екатерина Прекрасная',
+  'Хулио Иглесиас',
+  'Клим Ворошилов',
+  'Пайпер Перабо',
+  'Людвиг',
+  'Мила Каримова',
+  'Рональд Макдональд',
+  'Рома Зверев',
+  'Павлик Морозов',
+];
+export {NAMES, DESCRIPTIONS, MESSAGES};
